refactor: dedupe non-cached KDF params helper (OK-58640)
