+++ packages/kit-bg/src/services/ServiceHyperLiquid/ServiceHyperliquid.ts
@@ -6,7 +6,7 @@ import { isEqual, isNil, omit } from 'lodash';
 import pTimeout from 'p-timeout';
 
 import type { ICoreHyperLiquidAgentCredential } from '@onekeyhq/core/src/types';
-import { getPbkdf2KdfParamsForNonDbTx } from '@onekeyhq/shared/src/appCrypto/modules/pbkdf2';
+import { getPbkdf2KdfParamsForNonDbTxNoCache } from '@onekeyhq/shared/src/appCrypto/modules/pbkdf2';
 import {
   backgroundClass,
   backgroundMethod,
@@ -3616,15 +3616,10 @@ export default class ServiceHyperliquid extends ServiceBase {
           approveAgentResult.status === 'ok' &&
           approveAgentResult.response.type === 'default'
         ) {
-          const kdfParams = getPbkdf2KdfParamsForNonDbTx();
-          const webCryptoKdfParams =
-            kdfParams.kdfBackend === 'webcrypto'
-              ? { ...kdfParams, enablePbkdf2Cache: false }
-              : undefined;
           const encodedPrivateKey =
             await this.backgroundApi.servicePassword.encodeSensitiveText({
               text: privateKeyHex,
-              ...webCryptoKdfParams,
+              ...getPbkdf2KdfParamsForNonDbTxNoCache(),
             });
 
           const { credentialId } =
